LogQueue: use iv_list instead of GQueue to avoid excessive allocations
Previously each LogMessage was enqueued into a GQueue which allocates
a GList structure for each insertion, which causes another allocation
per LogMessage insertion.

This is changed to avoid this allocation:
  * instead of GQueue we use Linux kernel style linked lists
    (prev/next pointers are embedded in the struct)
  * since we need per-queue data, we embed the next/prev ptrs &
    queue related in a new structure called LogMessageQueueNode
  * we preallocate a set of QueueNodes in the LogMessage structure
  * we continously monitor how much nodes can be expected to be needed
    and adjust our pre-allocations accordingly.

This saves us one alloc/free per message per destination in the
fast path, and improves cache situation as well.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>