affile: only regular files are assumed to be always writable
The file destination avoids polling regular files as those are always
reported to be writable by the underlying kernel anyway. This is
however not true if a device file is being opened, like /dev/console.

A workaround was to use the pipe() driver, but that has other
consequences (like opening the pipes in read/write mode,
rather than write only).

Reported-By: Paul Krizak <paul.krizak@amd.com>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>