dbparser: fixed linking on Cygwin
Some issues remained in the previous patch, which is supposed to be fixed
by this one. Windows doesn't support -fPIC, and thus it is not supported
to link an archive file (.a) into a shared library.

This patch omits the static library in the case of the libdbparser.so
compilation and uses it only when compiling unit tests and pdbtool.

Signed-off-by: Corinna Vinschen <vinschen@redhat.com>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>