LogMessage: implement reference/ack caching
Instead of doing an atomic op every time a message is refed/unrefed/acked
make it possible to define code paths that "cache" refs that may not affect
the freeing of the LogMessage structure.

For example, the reader thread can be quite sure that the message itself
will not be freed until the last log_msg_unref() at the end of the
processing of the given message is invoked. Therefore all ref/unref
calls in this period can safely be done in a non-atomic manner.

The same applies to the writer side, which can also be sure that
a message is not going to be freed until the processing of that
message finishes in the log_writer_flush() function.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>