Handle non-epollable devices gracefully
It turned out that not just regular files are not expected to provide
epoll() semantics, but also /dev/null is lacking epoll support.

This patch adds autodetection whether a given fd is epollable, and reacts
accordingly. It also needs an updated ivykis (the one with iv_fd_pollable())
function, e.g. syslog-ng-3.3-ref5).

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>