resolve_sockaddr: avoid the use of g_strlcpy()
g_strlcpy() is an open coded implementation of strlcpy() in Glib where
it is not supplied, thus it is much slower than using the optimized string
routines on Linux for example (which lacks a system supplied
strlcpy implementation).

This patch reworks the resolve_sockaddr() function and uses memcpy() and
explicit NUL termination when needed.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>