templates: added syntactic sugar for using filter expressions in template functions
$(grep) and $(if) take a filter expression as argument, however such
expressions may require quoted strings. The default behaviour of the
parser in this case was to remove quotes.

This is fixed by not caring about quotes within parenthesises, so
$(grep) beomes usable with this syntax:

$(grep ('$FACILITY' == 'level4') $MSG)

This works since the parentheses around the filter expression is nicely
processed by the filter anyway, and the template function arg parser
ignores quotes in this case.

Reported-By: Valentijn Sessink <valentyn@blub.net>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>