LogWriter: added support for a new flag 'ignore-errors'
This flag make syslog-ng ignore write errors, since in special cases a
write error may indicate a temporary failure related to the current
message being written out.

Such an example is /proc/net/xt_recent/violators, where an ill formatted
message will cause a write error, but suspending the destination
because of that is not good.

The LogWriter will cancel the complete I/O operation, e.g. it is not
recommended to use flush_lines() or any other kind of write coalescing.

Reported-By: Valentijn Sessink <valentyn@blub.net>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>