templates: fixed @XX style msg_ref parsing
Function invocations (e.g. $(grep)) didn't get the complete
context as one extremal value for msg_ref (==0) which indicated that
the user didn't specify an msg_ref was not properly used in the code.

This basically meant that $(grep) only operated on a single element,
which is not very usable after all.

As it seems the unit test didn't cover all cases properly.

This patch fixes up all cases and adds unit test coverage.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>