config: Support setting a default suppress() globally.
While setting suppress() on a per-destination basis is flexible, when
one wants to have the same setting for all destinations by default,
it's more convenient to do that globally in the options block.

That is implemented herein: one can set a global suppress(), and it
will be the default for all destinations that support it, yet, can
still be overridden on a per-destination basis.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>