Display information about available modules
This patch implements --module-registry option for syslog-ng, and a changed
--version output. The first displays detailed information on available
modules, the second displays the list of available modules at all.

This way it is more easy to see what a given module adds to syslog-ng,
and also makes it possible to see what options were compiled as modules.

This patch removes some lines of the --version output that displayed
whether an optional functionality was compiled or not. (such as
mongodb or Process Accounting).


Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>