dbparser: implement inject-mode() option
In addition to reposting new messages to the internal() source, implement
inject-mode(pass-through) that sends synthetic messages further down
the pipeline. For example:

parser p_db { db-parser(); };

log { source(...);
	parser(p_db);
	log { ... };
};

In "internal" mode p_db() would repost generated messages to the internal()
source, possibly causing messages to go to a completely independent log path.

In "pass-through" mode, new messages are simply injected at the same
place as db-parser() is referenced.

This makes it somewhat easier to route correllated messages.

NOTE: this synthetic messages do not adhere to flow-control rules, so make
sure you leave some space in the destination FIFO in addition to the sum
of the windows of all sources feeding messages to the same destination.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>