fixed syslog() source framed message reception
If the frame length indicator is only partially read
and positioned at the end of the buffer, syslog-ng would assume that the
connection is closed, as it tries to read into a zero-length buffer.

Thanks for Andy for creating such a detailed description on how to
reproduce it here:

https://lists.balabit.hu/pipermail/syslog-ng/2011-March/016240.html

With that description it was easy to fix this nasty bug, which has been
sitting in syslog-ng for all eternity. :)

Reported-By: Andy Ruch <adruch2002@yahoo.com>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>