afsql: add support for SQL-schema defined default values in columns
syslog-ng always used an

INSERT INTO <table> (field1, field2...) VALUES ('val1', 'val2', ...)

statement, which causes problems if you want to use the auto-increment
feature present in various databases. This patch adds support for specfying
"default" for a field value in order to use the default as specified
for the column in question when defining the table structure.

The format is:

  sql(... columns('foo int default 1', 'bar') values(default))

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>