configure.in: Fix --enable-mixed-linking mode.
When mixed linking is enabled, the tools (pdbtool, in particular)
weren't linked against ivykis, which resulted in a link-time error,
due to libsyslog-ng using ivykis, but not being linked to it.

A similar problem was present in the linking of syslog-ng: -livykis
DID appear in @SYSLOGNG_DEPS_LIBS@, but the neccessary library paths
(when using the internal ivykis) did not.

This patch fixes both problems by introducing an IVYKIS_NO_LIBTOOL_LIBS
variable, to be used when the main binary is linked in mixed mode, and adding
IVYKIS_LIBS to TOOLS_DEPS_LIBS too.

Reported-By: Peter Czanik <czanik@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>