syslog() destination driver: fixed potential framing problem on contended connections
In case the sender is too slow processing messages, the syslog() driver
may get an SSL3_WRITE_PENDING error if TLS is enabled, or may generate
an invalid framing if it is not. Both will cause the connection to
be broken: in the SSL case by the client, in the non-SSL case by the server.

The reason is, that in some cases LogProtoFramedClient could attempt
2 write operations in the same poll iteration, even if the first one
already returned partial success. The 1st such operation is potentially
a frame header, the 2nd is the message payload. The problem happens if we
can't fully write the frame header.

Signed-off-by: Viktor Juhasz <jviktor@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>