inner-dest/inner-source plugins
This patch adds another kind of plugin to the syslog-ng core: plugins that
extend the functionality of an already existing source/destination driver.

For example, "suppress" the option to suppress duplicate messages could be
reimplemented as such a plugin and that'd make it immediately available
in sql() or mongodb() destinations, neither of which are currently
LogWriter based.

The original patch was contributed by Attila Nagy, but was significantly
changed by me.

The patch is quite large, as it contains some API rework, renaming
of some identifiers (log_driver_ prefix instead of log_drv_), and introduces
an additional level in the class hierarchy, causing a lot of extra "super"
to be added when referencing derived data.

Some notable changes in the API:
  * source drivers need to be derived from LogSrcDriver
  * destination drivers need to be derived from LogDestDriver
  * the init/deinit methods for those classes need to be invoked explicitly,
    which didn't even exist before
  * the queue method got an additional gpointer argument, to be used
    when plugins are in use. usual implementations for the queue method need
    not care about that argument and should automatically be set to NULL
    by the framework.

Signed-off-by: Attila Nagy <naat@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>