Fixed BSD timestamp parsing when the day is padded with spaces in front (fixes: D-03717)
Although the bugreport contained info that it'd be 'no-hostname' related, in
reality it isn't. The unit tests couldn't cover this case as the
timestamp that can typically contain such fields do not contain year information.

This patch fills this whole in the unit test and also fixes the
problem itself. It was caused by one of the performance improvement
patches, but doesn't affect PE 4.0 or older OSE versions.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>