logmsg.c: fixed refcache assertion
In single-threaded mode (and probably in multi-threaded mode
too, but less likely) it may happen that the number of cached unrefs
is zero (when the message is dropped). In that case logmsg_cached_refs
will be exactly the same as the value log_msg_refcache_start() initialized
to.

However this value is not accepted by the assertion, which causes syslog-ng
to abort.

Also added a long comment block that explains the rationale behind
the assertion.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>