LogReader: fixed a failed assertion case in log_source_queue()
LogReader could overflow its window, causing an assertion to
trigger in log_source_queue(). This was caused by the handling of
EPOLLERR cases.

At least on Linux, it may happen that EPOLLERR is set, while there's
still data in the socket buffer.  Thus in reaction to an EPOLLERR,
we could possibly send further messages without validating that
log_source_free_to_send() would allow us to, potentially
overflowing our window (and causing a failed assertion in
log_source_queue().

This can be reproduced by making sure that the window is regularly full
(for instance starting syslog-ng with -d to slow it down while it
prints messages to the console, running in non-threaded mode), and
interrupting the connection that send messages to syslog-ng. This
will reproduce the problem quite regularly.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>