fixed 100% CPU usage if SSL is enabled on a tcp() source
If SSL is being used, the socket is sometimes polled for readability,
sometimes for writability, depending on what the SSL lib is telling us
to wait for.

However, LogReader failed to deregister the direction which we were not
interested any more, which caused the output direction to be registered
even if that wouldn't help to proceed with the SSL handshake. This
caused a 100% CPU usage when the handshake was completed.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>