fixed possible segmentation fault at exit
This patch fixes a case where an I/O job might be submitted while the
current config is being deinitialized.

The root cause was that new jobs should not be allowed to start in
this phase, it could happen that two events happen in the same poll
iteration:
  * the last running I/O job exists
  * a new one is being submitted.

In this case, the last I/O job causes syslog-ng to proceed with the reload,
which resets main_loop_io_workers_quit to FALSE, allowing new jobs to start.

However, we are still not ready for that for at least one poll iteration.

The fix was to delay setting main_loop_io_workers_quit to FALSE for another
poll iteration using an iv_task.


Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>