value-pairs: A framework to filter key=value pairs.
This is the implementation of the value-pairs() framework discussed on
the list in the past.

The syntax is as follows:

    destination d_driver {
      driver(
        value-pairs(
          scope($SCOPE)
          exclude("S_*") # shell-style glob to exclude keys
          pair("my-own-key" "$SOURCE/$HOST")
          key("HOST_FROM")
        )
      );
    };

Of course, this is not only for destination drivers: there is support
for parsing arguments passed to template functions. The above set
would look like this in a template function:

$(function --scope $SCOPE --exclude S_* --pair my-own-key=$SOURCE/$HOST --key HOST_FROM)

The idea is that we select a scope (see later for the list of scopes
currently supported), from which we can exclude certain keys using a
shell-style glob pattern, or add custom key-value pairs of our own, or
add back pre-defined keys that we do want, but the scope doesn't have
by default.

The available scopes at the moment are:

    * rfc3164: Key-value pairs based on the old syslog protocol, also
      available as "core" and "base".
    * rfc5424: Key-value pairs based on the new syslog protocol, also
      available as "syslog_proto".
    * all_macros: All macros known to syslog-ng (this only includes the
      pre-defined macros!)
    * selected_macros: A smaller set of selected macros, which the
      developers deemed important (see the source for the list, for now).
    * nv_pairs: Name-value pairs, the results of SDATA or patterndb
      parsing.
    * everything: All of the above, combined into one big set.

On the driver side, supporting value-pairs is reasonably easy:

    driver_option
    	: value_pair_option {
    	  driver_dd_set_value_pairs(last_driver, $1);
    	}
    	| <etc...>
    	;

And in the driver itself, make a foreach callback:

    static gboolean
    driver_vp_foreach (const gchar *name, const gchar *value,
    		   gpointer user_data)
    {
      /* Do something here with the name-value pairs */

      return FALSE;
    }

And whenever you want to work with the set, call
value_pairs_foreach():

    value_pairs_foreach (self->value_pairs, driver_vp_foreach,
    		     msg, self->seq_num, USER_DATA);

And that's about it.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>