templates.c: refactor the template function API to make it easier to write non-simple functions
Simple functions are really simple, they usually do nothing but expand
their arguments as a template and perform some kind of transformation
on them.

However, in some cases a little bit more complex stuff needs to be done,
and in that case the API was really tiresome. Very long argument lists
needed to be passed around verbatim, method reuse was almost impossible,
etc.

This patch makes it a little bit more friendly, with the following changes:
  * introduction of LogTemplateInvokeArgs instead of long arg lists
  * make all functions reusable (prepare, eval, call, free):
    * state allocation is managed by the core
    * state free is just another method of the template function
  * eval function became optional5

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>