$(sanitize): add new template function useful to sanitize filenames
This function is useful for sanitizing macro values in filenames, so
that special characters (control characters, slashes).

The behaviour can also be customized by command line arguments as
described below:

$(sanitize [opts] $arg1 $arg2 ...)

Options:
  --ctrl-chars or -c             Filter control characters (default)
  --no-ctrl-chars or -C          Don't filter control characters
  --invalid-chars <set> or -i    Set of characters to be translated, default "/"
  --replacement <replace> or -r  Single character replacement for invalid chars.

An implementation was submitted by Andreas Piesk for syslog-ng 2.1.4, but
that was less general and caused all template expansion to be
sanitized. This function is selective, only works if you request it.

Arguments are concatenated using '/' so they would become separate
directory levels when used in filenames.

For example:

file("/var/log/$(sanitize $HOST $PROGRAM)/messages");

Would be roughly equivalent to:

/var/log/$HOST/$PROGRAM/messages

Except that if $HOST and $PROGRAM contains slashes, those will be
replaced by '_'.

Requested-By: Andreas Piesk <a.piesk@gmx.net>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>