value-pairs: split "nv-pairs" scope into "internal" and "public" values
By convention, syslog-ng prefixes all nv pair names it generates with
a dot (or without having a dot at all, like in legacy syslog
related macros). The idea being that the user (e.g. the configuration)
has a final word on what NV pairs she/he constitutes part of the event and
what not.

For instance, syslog-ng puts structured data in values
named '.SDATA.<id>.<param>', which makes it easy to use, but these fields
are almost never part of the event itself that the message payload carries.

With this patch, 3 nv-pair related scopes are introduced:
  * nv-pairs: contains all the NV pairs that have no '.' prefix
  * dot-nv-pairs: contains all NV pairs that have a '.' prefix
  * all-nv-pairs: contains all NV pairs

Also a new scope is introduced named "sdata" which contains only .SDATA
prefixed data.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>