fixed memory leak using macros in template
Always allocate memory for self->filename, so that it can always be
freed properly. The former method of trying to figure out when to free
it and when to ignore it since it's owned by something else didn't
quite work, and led to memory leaks.

This patch also changes AFFileDestWriter->filename to be a simple char
pointer instead of a GString.

Signed-off-by: Peter Gyorko <gyorkop@balabit.hu>
Signed-off-by: Viktor Juhasz <jviktor@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>