value-pairs: Move the key transformators to a separate file.
In order to make the code more readable, moved the key transformators
to a different file, and redesigned the way they work:

Now all transformators have a base class, that has a match, a
transform and a destroy function. Derived classes can easily add new
fields, and set their own implementations for the functions.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>