value-pairs: Update value_pairs_new_from_cmdline() to 3.4.
A while ago the template functions were changed so that their argv[0]
became the template function's name.

What this means for value-pairs() is that we no longer need to play
dirty tricks with its argv[] to please g_option_context_parse(), which
expected the zero-index argument to be a program name.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>