LogReader: Proper wakeup function handling for deinitialized LogReaders
The wakeup() callback can be called even if a LogReader is
deinitialized (when the Writer is deinitialized _after_ the reader, and
some messages are flushed out in the deinit path), in
which case we shouldn't trigger the schedule_wakeup event.

The original patch by Viktor Juhasz, but was completely
reworked by Balazs Scheidler.

Signed-off-by: Viktor Juhasz <jviktor@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>