template: make template compilation thread safe
Instead of doing template compilation on-demand, require the application
code to do it explicitly, at initialization, in order to avoid having to
acquire a lock during runtime.

This patch contains an incompatible API change, and requires updating
all call-sites.

The change makes it also possible to change template string of a constructed
LogTemplate class multiple times with a call to log_template_compile().

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>