[apphook] call msg_init first and call msg_shutdown last
In order to make sure that internal message functions work as early
as possible and as long as possible.

Also, set log_stderr to TRUE when msg_deinit() gets called, which means
that we'll try to write the error messages to stderr in case we'd still
emit internal messages after deinit.

Signed-off-by: NagyAttila <naat@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>