logwriter: use format_uint32 instead of sprintf
sprintf is too slow to be used on the fast path and a hand-coded
number formatting function was created, which is supposed to be replace
to sprintf() invocations. One spot was missed however and this patch
changes the code to use it in Log_writer_format_log() when an RFC5424
style message is to be formatted.

Signed-off-by: Viktor Juhasz <jviktor@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>