LogWriter: ignore throttle at our last attempt to flush all messages out
When syslog-ng stops, it tries one last time to flush all data from
its buffers towards the destination. Until now this was limited
by the throttle() setting, however it makes more sense to flush as
fast as possible, since if throttle kicks in the remaining messages
are lost for good.

Signed-off-by: Peter Gyorko <gyorkop@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>