pipe(): fixed a possible busy loop on AIX
AIX behaves strangely when polling a pipe(). It indicates writability
if a single byte is still available in the buffer, but then doesn't
take a partial buffer, and refuses our write() operation with EAGAIN.

This means that we think that the poll is writable, but at the end we fail
to write, syslog-ng gets into a busy loop.

The solution is to retry the write() operation with ever decreasing buffer
sizes until it succeeds.

Strictly speaking the AIX implementation is conforming, but I haven't
seen this behaviour elsewhere, so it is only enabled for AIX.

Signed-off-by: Attila Nagy <naat@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>