delegate flow-control early-ack decision to destination drivers
Sources always expect an acknowledgement for each message they produce.
This acknowledgement is usually generated when all the destinations have
finished their processing of the given message, unless flow-control
is disabled.

Earlier the LogMultiplexer object was responsible for ACKing in the name
of destinations where flow-control was not desired. This patch changes
that and delegates the early-ack (or break-ack) decision to destination
drivers, based on the information stored in the LogPathOptions structure,
where the user-requested mode of operation is stored.

This patch splits the "flow_control" member of the LogPathOptions structure
into two:

  - ack_needed: specifies whether the source still expects an ACK on the
    current path.
  - flow_control_requested: specifies whether the user has enabled
    flow control on the current path.

Based on these, the destination driver is able to make a decision when
to perform the ACK:

  - log_msg_break_ack(): is a new function call that can be used to generate
    the ACK and prepare for _sending_ the same message for further processing.
  - log_msg_ack(): is the good-old ack generation function that can be
    used when the processing is finally finished, and no processing is to
    be done.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>