[logwriter] flush the queue completely at reload/exit even quit is pending
When the configuration is reloaded or syslog-ng exits, one last attempt
to completely flush the queue is performed. In this mode, the flush thread
will not stop immediately when the quit signal is detected, but rather
it loops through queued entries as long as the destination is able to accept
them.

Since the same quit signal causes readers to exit, only a finite amount of
messages may be pending. Also, in case of network destinations, the network
window will probably fill up, and that also causes the loop to exit. File
destinations will be completely written however.

Signed-off-by: Viktor Juhasz <jviktor@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>