[center] this patch brings back the correct inheritance and propagation of flow control
The last flow-control processing change broke the case of flow-control with
embedded log statements. The original behaviour (prior breakage) was to
enable flow control if any of the log statement branches had
flags(flow-control) in it. The patch delegating the flow control
decision to the destinations (more precisely: b8cc9fe) broke that.

Signed-off-by: Juhasz Viktor <jviktor@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>