logwriter: fixed log_writer_init() if self->proto is already set
In case the LogWriter instance gets deinited with a self->proto set, the
upcoming init didn't take that into account, and unless the call site
called log_writer_reopen(), it didn't work.

This patch changes that, in case there's already a self->proto (e.g. the
writer is being reused from an old config and SIGHUP processing finishes)
LogWriter works as if the application also called log_writer_reopen().

If the application would want to do a reopen, it can do so as it pleases.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>