[stats] Make the statistics more thread safe
To ease the transition, the original thread safety patch on stats registration
required the code either to:

1) grab stats_lock(), OR
2) to run in the main thread.

The 2nd case turned out to be bogus, as worker threads might be
still running when the main thread operated on stats counters. This patch
changes this by always grabbing the stats_lock(), and removing the 2nd
case from the assertion that checked if locking was correct.

Signed-off-by: Viktor Juhasz <jviktor@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>