[file] Avoid the race condition between queuing and reap.
It is possible that the destination writer is reaped, but in the
same time a new log arrives for the destination writer
which log fits for the current reaped writer but the writer gets destroyed.

In case of non-macroed destination, self->single_writer becomes NULL
in this case. The solution is to protect the queuing mechanism with
self->lock.

Signed-off-by: Viktor Juhasz <jviktor@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>