[sql] send a signal to the database thread only in case when the queue is syncronised
because in other cases the database thread wake up unnecessary many times in case of threaded mode
And the sql stucked if the explicit commit is turned on and flush_lines and log_iw_size
and fetch_limit and log_fifo_size have the same value. The reason of the stuck was that
the queue become full but the database thread can't pop any messages form the queue.
So it can be occurred that the reader finished the io_job and suspended
and the database thread wait for the signal with 0 read message.

Signed-off-by: Viktor Juhasz <jviktor@balabit.hu>