program-override(): disable the effect of store-legacy-msghdr flag
During 3.1, the flag "store-legacy-msghdr" has become default, which means
that the $PROGRAM[$PID] portion of the original syslog message was restored
to the same format it was received as, instead of being reconstructed
from the parsed values.

However, when specifying program-override(), this behaviour was
not restored, e.g. even if the value of the $PROGRAM name-value pair
has changed, the originally stored format was used in files.

This was certainly not very intuitive. This patch changes that, if
$PROGRAM is set in any way (e.g. program-override(), rewrite rule, parser)
the effect of store-legacy-msghdr is disabled. That can be restored
by using an explict $LEGACY_MSGHDR macro in the destination template.

Cc: syslog-ng-stable@balabit.hu
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>