templates: Add a C_* family of timestamp macros.
The C_* family of timestamp macros work very similarly to R_
(received) and S_ (stamp), except C_ uses the current system time as
of the macro expansion.

The patch below uses cached_g_current_time() for performance reasons,
but in case the minor caching is undesired, it's easy to convert to
something else.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>