spoof-source: fixed memory leak & crash
spoof-source didn't properly lock the libnet context that was used to
craft the outgoing frame. This results in massive memory leaks and
crashes, as libnet stores the outgoing frame in its libnet_ctx structure.

NOTE: spoof-sourced frames are sent in the input thread, this could
also be solved by moving the functionality into the output thread (sparing
a couple of locks and better scalability on the read side too),
but that would've been a larger change.

Reported-By: Martin Holste <mcholste@gmail.com>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>