nv_table_clone(): fixed a memory leak by setting "borrowed" to FALSE
NVTable->borrowed indicates whether the memory area occupied by NVTable
itself is borrowed by an encompassing structure (such as the LogMessage).

nv_table_clone() clones an existing NVTable on a newly allocated region,
and since it is allocated on the heap on its own, "borrowed" should
be FALSE, however it was copied from the to-be-cloned structure verbatim.

Ensure it gets set to FALSE.

Cc: <syslog-ng-stable@balabit.hu>
Reported-by: Martin Holste <mcholste@gmail.com>
Reported-by: Gergely Nagy <algernon@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>