SDATA: include meta.sequenceId value in RFC5424 formatted message
The code to include meta.sequenceId was commented out some time ago, since
it modified the message right before sending it out on the wire, effectively
ruining the message model: this change propagated to parallel branches of
the processing pipeline where it shouldn't have.

This change fixes that, instead of modifying the message before formatting
it with the template, the SDATA payload is automatically extended without having
to store sequenceId in the message.

Signed-off-by: Viktor Juhasz <jviktor@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>