internal(): add locking on next_mark_target
This patch adds locking around the global variable "next_mark_target" to
ensure that the MARK message generation finds a consistent value there.

An alternative would be to change next_mark_target to be a simple int
instead of struct timespec, and set/get using atomic operations, but
this lock is only taken for local messages, thus I defer this
optimization until we actually measure it is worth it.

Signed-off-by: Attila Nagy <naat@balabit.hu>
Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>