nvtable: Handle errors from registry_alloc_handle & debug messages
When the length of a key is zero, do not pass it to
nv_registry_alloc_handle(), as it will fail. In these cases, and
whenever alloc_handle() returns 0, do not use the result for anything.

On one side, this should considerably reduce the amount of error
messages we receive, and the added debug logs will be more useful for
us in tracking down the root cause.

    NOTE: the original patch has been optimised somewhat to avoid the costy
    strlen() calls as this is in a fastpath...

Signed-off-by: Juhasz Viktor <jviktor@balabit.hu>