[logwriter] unified reopen handling the reopen timer is handled in log_writer
And if the reopen timer elapsed the LogWriter send a notification (NC_REOPEN_REQUEST)
that's the notification what the destination drivers have to handle.
Previously the reopen was handled different way in afsocket and affile.
This is more simple, because the difference is removed only the notification handlings
are differ.

Signed-off-by: NagyAttila <naat@balabit.hu>