crypto: autodetect the need to seed the PRNG
This patch changes the crypto.c module to perform an autodetection whether
the random number generator needs to be seeded.

RAND_loadfile() is triggered if the environment variable RANDFILE is set, or
when openssl itself is unable to detect how to seed the random
number generator.

This gets rid of the --seed command line option too (which is
kept for compatibility for now).

A warning is printed on stderr if everything fails.

Signed-off-by: Balazs Scheidler <bazsi@balabit.hu>